+++ schmellApp/package.json
@@ -37,6 +37,7 @@
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
     "@tsconfig/react-native": "^2.0.2",
+    "@types/crypto-js": "^4.1.1",
     "@types/jest": "^26.0.23",
     "@types/react-native": "^0.69.3",
     "@types/react-test-renderer": "^18.0.0",
