+++ schmellApp/src/screens/GamePlay/index.tsx
@@ -9,8 +9,8 @@ import {setQuestions} from '../../features/gameplay/gamePlaySlice';
 import {useAppDispatch} from '../../features/hooks';
 import {selectPlayers, selectQuestions} from '../../features/selectors';
 import {useAddPlayerToQuestionsQuery} from '../../services/apiService';
-import {modalShowType} from '../../typings/common';
-import {questionType} from '../../typings/question';
+import {modalShowType} from '../../types/common';
+import {questionType} from '../../types/question';
 import Carousel from './Carousel';
 import GameFooter from './GameFooter';
 import GameHeader from './GameHeader';
