+++ schmellApp/src/screens/Home/GameButton.tsx
@@ -4,8 +4,8 @@ import {Animated} from 'react-native';
 import {setSelectedGame} from '../../features/game/gameSlice';
 import {useAppDispatch} from '../../features/hooks';
 import {selectGameDetail} from '../../features/selectors';
-import {gameType} from '../../typings/game';
-import {HomeScreenNavigationProp} from '../../typings/navigation';
+import {gameType} from '../../types/game';
+import {HomeScreenNavigationProp} from '../../types/navigation';
 import {useSelector} from 'react-redux';
 import {useNavigation} from '@react-navigation/native';
 import SchmellButton from '../../components/Buttons/SchmellButton';
