+++ schmellApp/src/screens/Home/GameDetail.tsx
@@ -1,8 +1,8 @@
 import React, {FC, useState} from 'react';
 import {Animated, Image, Text, View} from 'react-native';
 import {Dispatch} from '@reduxjs/toolkit';
-import {gameType} from '../../typings/game';
-import {showDetailType} from '../../typings/settings';
+import {gameType} from '../../types/game';
+import {showDetailType} from '../../types/settings';
 import {postDetail} from '../../features/usersettings/userSettingSlice';
 import {setSelectedGame} from '../../features/game/gameSlice';
 import {useAppDispatch} from '../../features/hooks';
@@ -10,7 +10,7 @@ import {selectGameDetail, selectLanguage} from '../../features/selectors';
 import useLocale from '../../hooks/useLocale';
 import {useSelector} from 'react-redux';
 import {useNavigation} from '@react-navigation/native';
-import {HomeScreenNavigationProp} from '../../typings/navigation';
+import {HomeScreenNavigationProp} from '../../types/navigation';
 import homeStyle from './style';
 import Checkbox from '../../components/Forms/Checkbox';
 import SchmellButton from '../../components/Buttons/SchmellButton';
