+++ schmellApp/src/screens/Home/index.tsx
@@ -8,11 +8,11 @@ import homeStyle from './style';
 import {Dispatch} from '@reduxjs/toolkit';
 import {useAppDispatch} from '../../features/hooks';
 import {postSettings} from '../../features/usersettings/userSettingSlice';
-import {userSettings} from '../../typings/settings';
+import {userSettings} from '../../types/settings';
 import checkIfSettingsSet from '../../utils/checkIfSettingsSet';
 import {useSelector} from 'react-redux';
 import {selectUserStatus} from '../../features/selectors';
-import {lockPortrait} from '../../utils/orientationLocker';
+import {lockPortrait} from '../../utils/orientation';
 import {useSetTokensQuery} from '../../services/apiService';
 import RNUniqueId from '../../native/RNUniqueId';
 import encryptedStorageService from '../../services/encryptedStorageService';
