+++ schmellApp/src/types/api.ts
