+++ schmellApp/src/types/common.ts
