+++ schmellApp/src/types/navigation.ts
