+++ schmellApp/src/types/react-native-svg.d.ts
