+++ schmellApp/src/types/settings.ts
