+++ schmellApp/src/types/state.ts
