+++ schmellApp/src/types/week.ts
