--- schmellApp/src/utils/numberUtil.ts
@@ -1,3 +1,0 @@
-export function getRandomNumber(arraySize: number) {
-  return Math.floor(Math.random() * arraySize);
-}
