+++ schmellApp/src/utils/orientation.ts
